open_food_api/requirements.txt
Django==4.1.4
djangorestframework==3.12.4